Normalize importance to INTEGER 0-100 using 60/40 density/IDF weighting
Replace unbounded DOUBLE importance with a clamped 0-100 INTEGER score:
60% from S2 level-12 density (ln(1+count)/10, capped at 1) and 40% from
category IDF (idf_score/18, capped at 1). Both import scripts now require
density.parquet and category_idf.parquet (exit 1 if missing) instead of
degrading silently. Test fixtures updated to use varied INTEGER values in
the 0-100 range so type and non-uniformity assertions pass.

+++ tests/test_database.py
@@ -199,7 +199,7 @@ def test_trigram_index_detection(tmp_path):
             name VARCHAR,
             lat DOUBLE,
             lon DOUBLE,
-            importance DOUBLE
+            importance INTEGER
         )
     """)
     conn.close()
@@ -220,7 +220,7 @@ def test_trigram_index_detection_missing(tmp_path):
             dm_code VARCHAR,
             fsq_place_id VARCHAR,
             name VARCHAR,
-            importance DOUBLE
+            importance INTEGER
         )
     """)
     conn.close()
